Move statistics tests out.
